Set model.$state.isExpanded only when model.metadata.hasChildren is true
Row without children can not be expanded, therefore setting true to `isExpanded` to a row without any produces an illogical state.
This may lead (or leads) to bugs when the state of a row is used i.e. for styling.
This fix should set `model.$state.isExpanded` only when such row has children.